+++ packages/showcases/core/csf/components/forms/select.csf.js
@@ -1,6 +1,6 @@
 export const parameters = {
   actions: {
-    handles: ['mouseenter', 'click', 'focusin', 'focusout'],
+    handles: ['mouseenter', 'click', 'focusin', 'focusout', 'change'],
   },
   design: {
     type: 'figma',