+++ packages/sources/react/src/components/structure/VtmnDivider/VtmnDivider.tsx
@@ -2,6 +2,7 @@ import * as React from 'react';
 import clsx from 'clsx';
 import '@vtmn/css-divider/dist/index-with-vars.css';
 import { VtmnDividerOrientation, VtmnDividerTextPosition } from './types';
+import { objectValuesToString } from '../../../utils/object';
 
 export interface VtmnDividerProps
   extends React.ComponentPropsWithoutRef<'div'> {
@@ -33,7 +34,7 @@ export const VtmnDivider = ({
         `vtmn-divider_orientation--${orientation}`,
         className,
       )}
-      {...props}
+      {...objectValuesToString(props)}
       role="separator"
       aria-orientation={orientation}
     >