+++ packages/sources/svelte/src/components/index.js
@@ -5,16 +5,20 @@ export { default as VtmnDropdownItem } from './actions/VtmnDropdownItem/VtmnDrop
 export { default as VtmnLink } from './actions/VtmnLink/VtmnLink.svelte';
 
 // Forms
+export { default as VtmnSelect } from './forms/VtmnSelect/VtmnSelect.svelte';
 export { default as VtmnTextInput } from './forms/VtmnTextInput/VtmnTextInput.svelte';
 
 // Indicators
 export { default as VtmnBadge } from './indicators/VtmnBadge/VtmnBadge.svelte';
+export { default as VtmnLoader } from './indicators/VtmnLoader/VtmnLoader.svelte';
 export { default as VtmnPrice } from './indicators/VtmnPrice/VtmnPrice.svelte';
+export { default as VtmnProgressbar } from './indicators/VtmnProgressbar/VtmnProgressbar.svelte';
 export { default as VtmnRating } from './indicators/VtmnRating/VtmnRating.svelte';
 export { default as VtmnTag } from './indicators/VtmnTag/VtmnTag.svelte';
 
 // Navigation
 export { default as VtmnSearch } from './navigation/VtmnSearch/VtmnSearch.svelte';
+export { default as VtmnNavbar } from './navigation/VtmnNavbar/VtmnNavbar.svelte';
 
 // Overlays
 export { default as VtmnAlert } from './overlays/VtmnAlert/VtmnAlert.svelte';
@@ -25,10 +29,13 @@ export { default as VtmnSnackbar } from './overlays/VtmnSnackbar/VtmnSnackbar.sv
 export { vtmnSnackbarStore } from './overlays/VtmnSnackbar/vtmnSnackbarStore';
 export { default as VtmnToast } from './overlays/VtmnToast/VtmnToast.svelte';
 export { vtmnToastStore } from './overlays/VtmnToast/vtmnToastStore';
+export { default as VtmnTooltip } from './overlays/VtmnTooltip/VtmnTooltip.svelte';
 
 // Selection controls
 export { default as VtmnCheckbox } from './selection-controls/VtmnCheckbox/VtmnCheckbox.svelte';
 export { default as VtmnChip } from './selection-controls/VtmnChip/VtmnChip.svelte';
+export { default as VtmnQuantity } from './selection-controls/VtmnQuantity/VtmnQuantity.svelte';
+export { default as VtmnRadioButton } from './selection-controls/VtmnRadioButton/VtmnRadioButton.svelte';
 export { default as VtmnToggle } from './selection-controls/VtmnToggle/VtmnToggle.svelte';
 
 // Structure